pyproject.toml
[tool.poetry]
name = "lex-vector"
version = "0.1.0"
description = ""
authors = ["Jakub Chojnacki <james@marl.engineering>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
poethepoet = "^0.21.1"
chromadb = "^0.4.22"
pre-commit = "^3.6.0"
torch = "^2.1.2"
umap-learn = "^0.5.5"
matplotlib = "^3.8.2"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
sentence-transformers = "^2.2.2"
langchain = "^0.1.0"
pypdf = "^3.17.4"
tqdm = "^4.66.1"
openai = "^1.7.1"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.12.0+cu113 torchvision==0.13.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html"
